+++ src/content/DynamicList/index.js
@@ -1,0 +1,1 @@
+export { DynamicList } from './DynamicList';