+++ CHANGELOG.md
@@ -31,3 +31,5 @@
 
 #v1.0.0-alpha.2
 - Feature: Dialog
+- Feature: Disabled Action Button
+- Feature: Dynamic List